refactor(observability): split type tests for GH727
