+++ specs/GH727/product.md
@@ -6,9 +6,9 @@ GH-727 / #727
 
 ## 用户问题
 
-`origin/main@13ff2d1d` 仍有 23 个 tracked Rust 文件超过 U-16 的 800 行硬上限。当前最大文件是
-`src/core/analytics/reports.rs`，它是一个 860 行 analytics report module，其中 production report
-template/data/generator definitions 约 225 行，其余是 inline unit tests。
+`origin/main@b1e5b85f` 仍有 22 个 tracked Rust 文件超过 U-16 的 800 行硬上限。当前最大文件是
+`src/core/observability/types.rs`，它是一个 856 行 observability type module，其中 production metric/log/span
+definitions 约 246 行，其余是 inline unit tests。
 
 本轮目标继续执行完整的大文件解耦计划：每个 PR 仍然小而可审，但所有 tranche 都必须服从同一套
 架构边界，避免制造新的耦合、重导出混乱或行为漂移。
@@ -34,40 +34,40 @@ template/data/generator definitions 约 225 行，其余是 inline unit tests。
 
 ## 本 tranche 目标
 
-- 拆分 `src/core/analytics/reports.rs`，它当前 860 行，是 #727 当前最大的 tracked Rust 文件。
-- 保留 `core::analytics` 对 report 相关 public type 的 re-export 不变。
-- 将 root `reports.rs` 保持为 production report definition/generator file，并用 `#[path = "reports_tests.rs"] mod tests;` 委托测试。
-- 将原 inline tests 移动到 `src/core/analytics/reports_tests.rs`，不改变断言或 fixtures。
-- 因 production definitions 本身低于 800 行，本 tranche 不拆 analytics report public surface，避免无意义 facade churn。
+- 拆分 `src/core/observability/types.rs`，它当前 856 行，是 #727 当前最大的 tracked Rust 文件。
+- 保留 `core::observability` 对 observability type 的 re-export 不变。
+- 将 root `types.rs` 保持为 production metric/log/span type definition file，并用 `#[path = "types_tests.rs"] mod tests;` 委托测试。
+- 将原 inline tests 移动到 `src/core/observability/types_tests.rs`，不改变断言或 fixtures。
+- 因 production definitions 本身低于 800 行，本 tranche 不拆 observability type public surface，避免无意义 facade churn。
 - 所有新增或修改后的 Rust 文件低于 800 行。
 
 ## 非目标
 
-- 不修改 analytics report runtime behavior、template defaults、report ID generation、summary defaults、serialization fields 或 `core::analytics` re-export surface。
-- 不改变 `ReportTemplate`、`ReportSection`、`GeneratedReport`、`ReportSectionData`、`ChartData`、`DataPoint`、`ReportSummary` 的字段、derive、serde 行为或 test assertions。
-- 不在本 PR 中处理其余 22 个大文件。
+- 不修改 observability runtime behavior、logging/tracing/metrics modules、log-entry conversion behavior、serde fields 或 `core::observability` re-export surface。
+- 不改变 `MetricValue`、`ObservabilityLogRecord`、`LogLevel`、`TokenUsage`、`ErrorDetails`、`AlertCondition`、`AlertSeverity`、`AlertState`、`TraceSpan`、`SpanLog` 的字段、derive、serde 行为或 test assertions。
+- 不在本 PR 中处理其余 21 个大文件。
 - 不关闭 #727。
 
 ## Behavior Invariants
 
-1. `core::analytics::{ReportGenerator, ReportTemplate, ReportSection, ReportSectionType, ReportFormat, GeneratedReport, ReportSectionData, ChartData, DataPoint, ReportSummary}` remain available through the original analytics re-export path.
-2. Analytics report type fields, derives, defaults, serde attributes, and generator template behavior remain unchanged.
-3. The root `reports.rs` remains the production analytics report definition/generator file; tests are delegated only through `#[path = "reports_tests.rs"]`.
-4. Original inline tests remain under `core::analytics::reports::tests`.
-5. No analytics runtime, collector, engine, optimizer, or types modules are changed.
+1. `core::observability::{MetricValue, ObservabilityLogRecord, LogEntry, LogLevel, TokenUsage, ErrorDetails, AlertCondition, AlertSeverity, AlertState, TraceSpan, SpanLog}` remain available through the original observability re-export path.
+2. Observability type fields, derives, serde attributes, and `LogEntry` conversion behavior remain unchanged.
+3. The root `types.rs` remains the production observability type definition file; tests are delegated only through `#[path = "types_tests.rs"]`.
+4. Original inline tests remain under `core::observability::types::tests`.
+5. No observability runtime, logging, metrics, redaction, destinations, histogram, or tracing modules are changed.
 6. Every touched Rust file must be below U-16's 800-line ceiling.
-7. `cargo test core::analytics::reports --lib --all-features` must pass.
+7. `cargo test core::observability::types --lib --all-features` must pass.
 
 ## 验收标准
 
-- [ ] `src/core/analytics/reports.rs` keeps production analytics report definitions and delegates tests with `#[path = "reports_tests.rs"] mod tests;`。
-- [ ] Original inline analytics report tests move to `src/core/analytics/reports_tests.rs` without assertion changes。
-- [ ] Public analytics report exports and behavior remain unchanged。
-- [ ] Both touched analytics report files are below U-16's 800-line ceiling。
-- [ ] Focused analytics report test suite 通过。
+- [ ] `src/core/observability/types.rs` keeps production observability type definitions and delegates tests with `#[path = "types_tests.rs"] mod tests;`。
+- [ ] Original inline observability type tests move to `src/core/observability/types_tests.rs` without assertion changes。
+- [ ] Public observability type exports and behavior remain unchanged。
+- [ ] Both touched observability type files are below U-16's 800-line ceiling。
+- [ ] Focused observability type test suite 通过。
 - [ ] `cargo fmt --all -- --check`、`cargo check --lib --all-features`、`cargo check --all-features --locked` 和 `cargo check` 通过。
 - [ ] PR body 明确该 PR 是 #727 的 partial tranche，使用 `Refs #727`，不自动关闭 tracker issue。
 
 ## 发布说明
 
-No runtime behavior change. This is an analytics reports test extraction for U-16 compliance.
+No runtime behavior change. This is an observability types test extraction for U-16 compliance.
